test: move JobIT test to ResolutionActionIT
